refactor: update imports and improve consistency across shared modules and tests
- Replaced integer status codes with string constants ("SUCCESS") for better clarity in HBAR transfer integration tests.
- Enhanced token tools and query configurations in test setup to align with updated plugin structure.
- Fixed `maxSupply` handling in `hedera-parameter-normaliser` to ensure `.toNumber()` is correctly applied for consistency.
